Change the way supported locales is set
The test suite compilation seems to end in error when using "of" method from
Set.

Change the parameter initialization to use HashSet so that it works with Java
older than 9 too.

The method was introduced in Java 9, but we are already using Java 11. So
it's a mystery why the "of"-method is missing.